+++ b2sdk/_v3/__init__.py
@@ -61,7 +61,6 @@ from b2sdk.utils import (
     hex_sha1_of_stream,
     hex_sha1_of_bytes,
     hex_sha1_of_file,
-    TempDir,
     IncrementalHexDigester,
 )